S8_Memory/S8_requirements.txt
langchain-core
langchain-openai
langchain-community
python-dotenv
sqlalchemy
streamlit